+++ querybuilder/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class QueryBuilderConfig(AppConfig):
+    name = 'querybuilder'
+    verbose_name = 'Django Query Builder'