+++ run_tests.py
@@ -8,10 +8,9 @@ import django
 
 from settings import configure_settings
 
-
 # Configure the default settings
 configure_settings()
-if django.VERSION[1] == 7:
+if django.VERSION[1] >= 7:
     django.setup()
 
 # Django nose must be imported here since it depends on the settings being configured